also fix superclass

+++ custom/ilsgateway/tests/handlers/utils.py
@@ -99,7 +99,8 @@ class ILSTestScript(TestScript):
     @classmethod
     def tearDownClass(cls):
         delete_domain_phone_numbers(TEST_DOMAIN)
-        cls.sms_backend_mapping.delete()
+        if cls.sms_backend_mapping.id is not None:
+            cls.sms_backend_mapping.delete()
         cls.sms_backend.delete()
         CommCareUser.get_by_username('stella').delete()
         CommCareUser.get_by_username('bella').delete()
